Separated the tests out
