+++ src/dotnet/CarbonAware.WebApi.Tests/integrationTests/SciScoreControllerTests.cs
@@ -1,0 +1,94 @@
+namespace CarbonAware.WepApi.IntegrationTests;
+
+using System;
+using System.Collections.Generic;
+using System.Net;
+using CarbonAware.Model;
+using CarbonAware.WebApi.Controllers;
+using Microsoft.AspNetCore.Mvc.Testing;
+using NUnit.Framework;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using System.Net.Http.Json;
+using Moq;
+using System.Net.Http.Headers;
+
+/// <summary>
+/// Tests that the Web API controller handles and packages various responses from a plugin properly 
+/// including empty responses and exceptions.
+/// </summary>
+[TestFixture]
+public class SciScoreControllerTests
+{
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+	private APIWebApplicationFactory _factory;
+	private HttpClient _client;
+#pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
+    [OneTimeSetUp]
+    public void Setup()
+    {
+        _factory = new APIWebApplicationFactory();
+        _client = _factory.CreateClient();
+    }
+
+
+    [Test]
+    public async Task SCI_WithValidData_ReturnsContent()
+    {
+        //Construct body object and then serialize it with JSONSerializer
+        object body = new
+        {
+            location = new
+            {
+                locationType = "CloudProvider",
+                providerName = "Azure",
+                regionName = "uswest"
+            },
+            timeInterval = "2007-03-01T13:00:00Z/2007-03-01T15:30:00Z"
+        };
+
+        var jsonBody = JsonSerializer.Serialize(body);
+        StringContent _content = new StringContent(jsonBody);
+
+        var mediaType = new MediaTypeHeaderValue("application/json");
+        _content.Headers.ContentType = mediaType;
+
+        var result = await _client.PostAsync("/sci-scores/marginal-carbon-intensity", _content);
+        var resultContent = await result.Content.ReadAsStringAsync();
+
+        Assert.That(result, Is.Not.Null);
+        Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.OK));
+        Assert.That(resultContent, Is.Not.Null);
+    }
+
+    [Test]
+    public async Task SCI_WithInvalidData_ReturnsBadRequest()
+    {
+        object body = new
+        {
+            location = new { },
+            timeInterval = "2007-03-01T13:00:00Z/2007-03-01T15:30:00Z"
+        };
+
+        var jsonBody = JsonSerializer.Serialize(body);
+        StringContent _content = new StringContent(jsonBody);
+
+        var mediaType = new MediaTypeHeaderValue("application/json");
+        _content.Headers.ContentType = mediaType;
+
+        var result = await _client.PostAsync("/sci-scores/marginal-carbon-intensity", _content);
+        var resultContent = await result.Content.ReadAsStringAsync();
+
+        Assert.That(result, Is.Not.Null);
+        Assert.That(result.StatusCode, Is.EqualTo(HttpStatusCode.BadRequest));
+    }
+
+
+    [OneTimeTearDown]
+    public void TearDown()
+    {
+        _client.Dispose();
+        _factory.Dispose();
+    }
+}
