--- src/dotnet/CarbonAware.WebApi.Tests/integrationTests/SciScoreIntegrationTests.cs
@@ -1,32 +1,0 @@
-namespace CarbonAware.WebApi.IntegrationTests;
-
-using System.Collections.Generic;
-using System.Net;
-using Microsoft.AspNetCore.Mvc.Testing;
-using NUnit.Framework;
-
-[TestFixture]
-public class SciScoreIntegrationTests
-{
-    [Setup]
-    public void Setup()
-    {
-        await using var application = new WebApplicationFactory<Program>();
-        var client = application.CreateClient(); // HttpClient
-
-    }
-    [Test]
-    public async Task Get_CarbonAwareController()
-    {
-        var response = await client.GetAsync("/emissions/bylocations/best");
-        AssertEqual(response.HttpStatusCode, HttpStatusCode.OK);
-    }
-
-    [Test]
-    public async Task Post_MarginalCarbonIntensity(){
-        // var requestContent = 
-        var response = await client.PostAsync("/sci-scores/marginal-carbon-intensity");
-        AssertEqual(response.HttpStatusCode, HttpStatusCode.OK);
-    }
-
-}
